Align simple spool cleanup WhatIf and restart behavior

+++ PowerShell Script/V7/Printer/Restart.Spool.DeletePrinterQSimple.ps1
@@ -1,5 +1,4 @@
 #Requires -Version 7.0
-#Requires -RunAsAdministrator
 
 [CmdletBinding(SupportsShouldProcess)]
 param()
@@ -7,28 +6,52 @@ param()
 Set-StrictMode -Version 3.0
 $ErrorActionPreference = 'Stop'
 
+$RequireAdmin = $true
+$IsAdministrator = ([Security.Principal.WindowsPrincipal] [Security.Principal.WindowsIdentity]::GetCurrent()).IsInRole([Security.Principal.WindowsBuiltInRole]::Administrator)
 $ScriptConfig = @{
-    ServiceName    = 'Spooler'
-    SpoolDirectory = Join-Path -Path $env:SystemRoot -ChildPath 'System32\spool\PRINTERS'
+    ServiceName       = 'Spooler'
+    SpoolDirectory    = Join-Path -Path $env:SystemRoot -ChildPath 'System32\spool\PRINTERS'
+    AllowedExtensions = @('.spl', '.shd')
 }
 
 function Invoke-ClearPrintQueueSimple {
     [CmdletBinding(SupportsShouldProcess)]
     param(
         [Parameter(Mandatory)]
+        [bool]$RequireAdmin,
+
+        [Parameter(Mandatory)]
+        [bool]$IsAdministrator,
+
+        [Parameter(Mandatory)]
         [string]$ServiceName,
 
         [Parameter(Mandatory)]
-        [string]$SpoolDirectory
+        [string]$SpoolDirectory,
+
+        [Parameter(Mandatory)]
+        [string[]]$AllowedExtensions
     )
 
-    Stop-Service -Name $ServiceName -Force -ErrorAction Stop
+    if ($RequireAdmin -and -not $WhatIfPreference -and -not $IsAdministrator) {
+        throw 'Run this script in an elevated PowerShell 7 session.'
+    }
+
+    $service = Get-Service -Name $ServiceName -ErrorAction Stop
+    $serviceWasRunning = $service.Status -eq [System.ServiceProcess.ServiceControllerStatus]::Running
+    $serviceWasStopped = $false
 
     try {
         $deletedCount = 0
+
+        if ($serviceWasRunning -and $PSCmdlet.ShouldProcess($ServiceName, 'Stop service')) {
+            Stop-Service -Name $ServiceName -Force -ErrorAction Stop
+            $serviceWasStopped = $true
+        }
+
         $files = @(
             Get-ChildItem -LiteralPath $SpoolDirectory -File -ErrorAction SilentlyContinue |
-                Where-Object { $_.Extension -in '.spl', '.shd' }
+                Where-Object { $AllowedExtensions -contains $_.Extension.ToLowerInvariant() }
         )
 
         foreach ($file in $files) {
@@ -44,14 +67,19 @@ function Invoke-ClearPrintQueueSimple {
         }
     }
     finally {
-        Start-Service -Name $ServiceName -ErrorAction Stop
+        if ($serviceWasStopped -and $PSCmdlet.ShouldProcess($ServiceName, 'Start service')) {
+            Start-Service -Name $ServiceName -ErrorAction Stop
+        }
     }
 }
 
 try {
     Invoke-ClearPrintQueueSimple `
+        -RequireAdmin $RequireAdmin `
+        -IsAdministrator $IsAdministrator `
         -ServiceName $ScriptConfig.ServiceName `
-        -SpoolDirectory $ScriptConfig.SpoolDirectory
+        -SpoolDirectory $ScriptConfig.SpoolDirectory `
+        -AllowedExtensions $ScriptConfig.AllowedExtensions
 }
 catch {
     Write-Error $_.Exception.Message
